+++ src/main/java/HuellaDeCarbono/CalculoDeDistancias/Distancia.java
@@ -1,10 +1,10 @@
 package HuellaDeCarbono.CalculoDeDistancias;
 
 public class Distancia {
-    private float valor;
+    private Double valor;
     private String unidad;
 
-    public float getValor() {return valor;}
+    public Double getValor() {return valor;}
     public String getUnidad() {return unidad;}
 
 }
